Delete bespoke Codpiece bounds
Replace unbounded interactive Codpiece traversal with a measured deadline, retain only conservative modifier-cache safety, and preserve exhaustive verification.
