+++ src/data/maximizer-help.html
@@ -18,7 +18,7 @@ Any numeric modifier keyword can be followed by one or both of these special key
 <br>Note that the limit keywords won't quite work as expected for a modifier that you're trying to minimize.
 <br>If <b>min</b> or <b>max</b> is specified at the start of the expression, it applies to the total score (the sum of each modifier value times its weight).  A global <b>max</b> may allow equipment maximization to finish faster, since no further combinations will be considered once the specified value is reached.
 <br>The preference "maximizerCombinationLimit" (set with "set maximizerCombinationLimit = 100000" in CLi) constrains the number of combinations Maximizer will consider. Depending where it is set, this may mean you don't get the best combination, but you may well get a good enough one faster. Depending which equipment you can use, you may find a higher value than 100000 is needed.
-<br>The preference "maximizerSearchTimeLimit" limits equipment search time in milliseconds (2000 by default, 0 for no limit). When the limit is reached, Maximizer returns the best legal result found so far without guaranteeing optimality.
+<br>The preference "maximizerSearchTimeLimit" limits equipment search time in milliseconds (500 by default, 0 for no limit). When the limit is reached, Maximizer returns the best legal result found so far without guaranteeing optimality.
 <h3>Other Modifiers</h3>
 Boolean modifiers can also be used as keywords.  With positive weight, the modifier is required to be true; with negative weight, it is required to be false.  There is one shortcut available: <b>sea</b> requires both Adventure Underwater and Underwater Familiar to be true.
 <p>
